--- js/get-post.js
@@ -1,5 +1,0 @@
-function getPost(posts, id) {
-  return posts.filter((post) => post.id === id)[0];
-}
-
-export { getPost };
