--- js/show-inline-note.js
@@ -1,7 +1,0 @@
-function showInlineNote(text) {
-  const note = document.createElement('div');
-  note.classList.add('inline-note');
-  note.textContent = text;
-}
-
-return showInline
